requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
email-validator
python-dotenv